Update Person.java
Updated to Integer

+++ spring-jpa/src/main/java/com/ssd/jpa/entity/Person.java
@@ -14,7 +14,7 @@ public class Person {
 	private String add;
 	
 	private Pancard pan;
-	private Long mobileNo;
+	private Integer mobileNo;
 	
 
 }
